+++ src/stores/Frame/Region/RegionSetStore.ts
@@ -143,8 +143,8 @@ export class RegionSetStore {
         return this.addRegion(points, 0, CARTA.RegionType.ANNRULER, temporary, true);
     };
 
-    @action addExistingRegion = (points: Point2D[], rotation: number, regionType: CARTA.RegionType, regionId: number, name: string, color: string, lineWidth: number, dashes: number[], annotationStyles?: any) => {
-        const region = this.addRegion(points, rotation, regionType, true, false, regionId, name);
+    @action addExistingRegion = (points: Point2D[], rotation: number, regionType: CARTA.RegionType, regionId: number, name: string, color: string, lineWidth: number, dashes: number[], temporary = true, annotationStyles?: any) => {
+        const region = this.addRegion(points, rotation, regionType, temporary, annotationStyles, regionId, name);
         // additional imported style properties;
         if (color) {
             region.color = color;