+++ CHANGELOG.md
@@ -25,6 +25,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Added the ability of changing to a new directory by entering a path ([#609](https://github.com/CARTAvis/carta-frontend/issues/609)).
 * Added supports for customizing histogram calculations ([#1488](https://github.com/CARTAvis/carta-frontend/issues/1488)).
 * Added pv image preview feature ([#1561](https://github.com/CARTAvis/carta-frontend/issues/1561)).
+* Added support for saving and restoring workspaces ([#1272](https://github.com/CARTAvis/carta-frontend/issues/1272)). Initial support is limited to restoring open images, render/contour/overlay configs and regions. 
+
 ### Fixed
 * Fixed the issue of annoying text input fields ([#1906](https://github.com/CARTAvis/carta-frontend/issues/1906)).
 * Fixed the issues of copying the Session URL in the macOS Electron and Linux AppImage versions ([#2102](https://github.com/CARTAvis/carta-frontend/issues/2102), [#2108](https://github.com/CARTAvis/carta-frontend/issues/2108)).